requirements.txt
asgiref==3.5.0
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
Django==3.2.13
django-crispy-forms==1.14.0
django-filter==21.1
Pillow==8.2.0
pytz==2022.1
sqlparse==0.4.2
